+++ Transacciones/eliminarUsuario.php
@@ -1,15 +1,17 @@
-<?php 
-require '../Conection/cn.php';
+<?php
+if ($_SERVER['$_POST']) {
+    require '../Conection/cn.php';
 
-$id = $db->real_escape_string($_POST['id']);
+    $id = $db->real_escape_string($_POST['id']);
 
-$query = "DELETE FROM usuarios WHERE id_usuario = $id ";
+    $query = "DELETE FROM usuarios WHERE id_usuario = $id ";
 
-if ($db->query($query)) {
-    header('Location: /soloarte/Paginas/Admin/Usuarios.php?resul=3');
+    if ($db->query($query)) {
+        header('Location: /soloarte/Paginas/Admin/Usuarios.php?resul=3');
+    } else {
+        echo 'Error al Eliminar';
+        exit;
+    }
 } else {
-    echo 'Error al Eliminar';
-    exit;
+    header('Location: /soloarte');
 }
-
-?>
